Move TestCommands from from sandbox-on-x-test-lib to test-common
