+++ ledger-test-tool/ledger-api-tests/infrastructure/BUILD.bazel
@@ -29,13 +29,13 @@ load("//ledger-test-tool/ledger-api-tests:conformance.bzl", "testtool_lf_version
             "//ledger/ledger-api-errors",
             "//libs-scala/test-evidence/tag:test-evidence-tag",
             "//test-common:dar-files-%s-lib" % lf_version,
-            "//test-common:test-common-%s" % lf_version,
             "//libs-scala/build-info",
             "//libs-scala/contextualized-logging",
             "//libs-scala/resources",
             "//libs-scala/resources-akka",
             "//libs-scala/resources-grpc",
             "//libs-scala/timer-utils",
+            "//test-common",
             "@maven//:com_google_api_grpc_proto_google_common_protos",
             "@maven//:com_typesafe_config",
             "@maven//:io_netty_netty_common",
