+++ ledger/sandbox-on-x/src/test/lib/scala/com/daml/platform/sandbox/services/SubmitAndWaitMultiPartyDummyCommand.scala
@@ -7,6 +7,7 @@ import java.util.UUID
 
 import com.daml.ledger.api.v1.command_service.{CommandServiceGrpc, SubmitAndWaitRequest}
 import com.daml.platform.sandbox.auth.ServiceCallAuthTests
+import com.daml.platform.testing.TestCommands
 import com.google.protobuf.empty.Empty
 
 import scala.concurrent.Future
