+++ release/artifacts.yaml
@@ -203,11 +203,7 @@
   type: jar-scala
 - target: //test-common:semantic-tests-1.dev.scala
   type: jar-scala
-- target: //test-common:test-common-1.14
-  type: jar-scala
-- target: //test-common:test-common-1.15
-  type: jar-scala
-- target: //test-common:test-common-1.dev
+- target: //test-common:test-common
   type: jar-scala
 - target: //test-common:carbonv1-tests-1.15.scala
   type: jar-scala
