+++ test-common/src/main/scala/com/daml/TestCommands.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2023 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.platform.sandbox.services
+package com.daml.platform.testing
 
 import java.io.File
 import java.util
